Update vyatta-update-crontab.pl
The cronjob being created when an hourly or daily interval is specified will run every minute during that hour or day because there's no preceding value that's defined. For the hourly one to work, it must have a minute defined. For the daily one to work, it must have a minute and hour defined. Proposed setting the values to "0" for the non-specified variables when using intervals.